Change project paths to use /projects/{user}/{slug} format

+++ app/Http/Controllers/ProjectController.php
@@ -4,10 +4,13 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Validation\Rule;
 use App\Project;
 use Redirect;
 use Response;
 use App\Http\helpers;
+use App\User;
+use Validator;
 
 class ProjectController extends Controller
 {
@@ -17,9 +20,16 @@ class ProjectController extends Controller
     return view('projects.index', compact('projects'));
   }
 
-  public function indexOwned()
+  public function my(Request $request)
   {
-      $user = Auth::user();
+      $user = $request->user();
+
+      return redirect()->route('projects.user', ['user' => $user]);
+  }
+
+
+  public function indexOwned(User $user)
+  {
       $projects = Project::where('owner_id', $user->id)->get();
       return view('projects.owned', compact('projects', 'user'));
   }
@@ -41,9 +51,14 @@ class ProjectController extends Controller
    * @param  int  $id
    * @return Response
    */
-  public function show(Project $project){
-    return view('projects.show', compact('project'));
+  public function show(User $user, $slug){
+      $project = Project::where(['owner_id' => $user->id, 'slug' => $slug])->get()->first();
+      if(!empty($project))
+      {
+          return view('projects.show', compact('project'));
+      }
 
+      abort(404);
   }
 
   /**
@@ -52,16 +67,41 @@ class ProjectController extends Controller
    * @param  int  $id
    * @return Response
    */
-  public function edit(Project $project)
+  public function edit(User $user, $slug)
   {
-    return view('projects.edit', compact('project'));
+      $project = Project::where(['owner_id' => $user->id, 'slug' => $slug])->get()->first();
+
+      if(!empty($project)){
+          return view('projects.edit', compact('project'));
+      }
+
+      abort(404);
+
   }
 
   public function store(Request $request)
   {
-      $this->validate($request, [
-          'name' => 'required|unique:projects|max:255',
-      ]);
+      $user = $request->user();
+
+      $rules = [
+          'name' => [
+              'required',
+              'max:255',
+              Rule::unique('projects')->where(function ($query) use ($user) {
+                  $query->where('owner_id', $user->id);
+              })
+          ]
+      ];
+
+      $messages = [
+          'name.unique' => 'It looks like you already have a project named "' . $request->input('name') . '". Try something else!',
+      ];
+      $validator = Validator::make($request->all(), $rules, $messages);
+
+      if($validator->fails()){
+          return redirect()->back()->withErrors($validator)->withInput();
+      }
+
 
     $project = new Project;
     $project->name = $request->input('name');
@@ -69,48 +109,49 @@ class ProjectController extends Controller
     $slug = str_slug($project->name, '-');
     $new_slug = $slug;
 
-    $exists = Project::where(['slug' => $slug])->count();
+    $exists = Project::where(['slug' => $slug, 'owner_id' => $user->id])->count();
     $append = 1;
     while($exists){
         $new_slug = $slug . '-' . $append;
         $append++;
-        $exists = Project::where(['slug' => $new_slug])->count();
+        $exists = Project::where(['slug' => $new_slug, 'owner_id' => $user->id])->count();
     }
 
     $project->slug = $new_slug;
-    $project->owner_id = $request->user()->id;
+    $project->owner_id = $user->id;
     $project->save();
 
     return redirect()->back()->with('message', 'Created new project ' . $project->name . '!');
   }
 
-  public function update(Request $request, Project $project)
+  public function destroy(Request $request, User $user, $slug)
   {
-    $input = array_except($request->all(), '_method');
-    $input['slug'] = !empty($input['name']) ? str_slug($input['name']) : $project->slug;
-    $project->update($input);
+      $project = Project::where(['owner_id' => $user->id, 'slug' => $slug])->get()->first();
 
-    return Redirect::route('projects.show', $project->slug)->with('message', 'Project updated.');
+      if(!empty($project)){
+          $this->validate($request, [
+              'project-name-confirm' => 'required|in:' . $project->fullPath(),
+          ]);
+
+        $project->delete();
+
+        return Redirect::route('projects.index')->with('message', 'Deleted project ' . $project->name . '!');
+    }
+    abort(404);
+
   }
 
-  public function destroy(Request $request, Project $project)
-  {
-      $this->validate($request, [
-          'project-name-confirm' => 'required|in:' . $project->fullPath(),
-      ]);
+  public function getTags(Request $request, User $user, $slug){
+      $project = Project::where(['owner_id' => $user->id, 'slug' => $slug])->get()->first();
 
-    $project->delete();
+      if(!empty($project)){
+          if($request->ajax()){
+              $tags = $project->tags()->get();
+              return Response::json($tags);
 
-    return Redirect::route('projects.index')->with('message', 'Deleted project ' . $project->name . '!');
-  }
-
-  public function getTags(Request $request, Project $project){
-      if($request->ajax()){
-          $tags = $project->tags()->get();
-          return Response::json($tags);
-
+          }
+          return view('projects.tags', compact('project'));
       }
-
-      return view('projects.tags', compact('project'));
+      abort(404);
   }
 }
